Refactored old APIs on SnappyContext to a different implicits. Changed tests accordingly.
