+++ snappy-tools/src/test/scala/io/snappydata/app/streaming/kafka/KafkaConsumer.scala
@@ -16,9 +16,6 @@
  */
 package io.snappydata.app.streaming.kafka
 
-/**
- * Created by ymahajan on 28/10/15.
- */
 
 import org.apache.spark._
 import org.apache.spark.rdd.RDD
@@ -26,6 +23,7 @@ import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.streaming.SnappyStreamingContext
 import org.apache.spark.sql.{Row, SaveMode, SnappyContext}
 import org.apache.spark.streaming._
+import org.apache.spark.sql.devapi._
 
 
 object KafkaConsumer {
